% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{active_burst<-}
\alias{active_burst<-}
\title{change active burst}
\usage{
active_burst(burst) <- value
}
\description{
change active burst
}
\examples{
 data(raccoon_data)
 burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
 mb1 <- make_multi_burst(burst_list=burstz, active_burst=c('id','month'))
 active_burst(mb1)
 active_burst(mb1) <- c('id')
 active_burst(mb1)
}
