% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{active_burst}
\alias{active_burst}
\alias{active_burst<-}
\title{Access the active_burst value}
\usage{
active_burst(burst)

active_burst(x, ...) <- value
}
\arguments{
\item{burst}{a multi_burst}

\item{value}{character vector of the burst names to make active}
}
\description{
The active burst is the combination of bursts that group the data sets.
The active_bursts are essentially a paste(names_of_bursts, sep = '_') grouping variable.
}
\examples{
raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
raccoon_data$acquisition_time <- as.POSIXct(raccoon_data$acquisition_time, 'EST')
 burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
 mb1 <- make_multi_burst(x=burstz, active_burst=c('id','month'))

 # see the current active burst
 active_burst(mb1)

 # change the active burst
 active_burst(mb1) <- 'id'

 # Using a full data set
 my_track <- as_sftrack(raccoon_data, time = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst = burstz)

 summary(my_track)

 # change active_burst
 active_burst(my_track$burst) <- 'id'

 summary(my_track)
#'
}
