% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{active_burst<-}
\alias{active_burst<-}
\alias{active_burst_replace}
\alias{active_burst<-.sftrack}
\alias{active_burst<-.sftraj}
\alias{active_burst<-.multi_burst}
\alias{active_burst<-.ind_burst}
\title{Set new active burst}
\usage{
active_burst(x) <- value

\method{active_burst}{sftrack}(x) <- value

\method{active_burst}{sftraj}(x) <- value

\method{active_burst}{multi_burst}(x) <- value

\method{active_burst}{ind_burst}(x) <- value
}
\arguments{
\item{value}{character vector of the burst names to make active}
}
\description{
Set new active burst
}
