% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tj.R
\name{active_group}
\alias{active_group}
\title{Access the active_group value}
\usage{
active_group(x)
}
\arguments{
\item{x}{a c_grouping}
}
\description{
The active group is the combination of group names to group the data sets.
The active_group acts essentially like a paste(names_of_groups, sep = '_') grouping variable.
}
\examples{

# Make a c_grouping
group <- list(id = raccoons$animal_id, month = as.POSIXlt(raccoons$timestamp)$mon +1)
mb1 <- make_c_grouping(x = group, active_group = c("id", "month"))

# see the current active group
active_group(mb1)

# change the active group
active_group(mb1) <- "id"

# Using a full data set
group <- list(id = raccoons$animal_id, month = as.POSIXlt(raccoons$timestamp)$mon+1)
my_track <- as_sftrack(raccoons,
  time = "timestamp",
  error = NA, coords = c("longitude", "latitude"),
  group = group
)

summary(my_track)

# change active group
active_group(my_track$sft_group) <- "id"

summary(my_track)
}
