% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{active_group<-}
\alias{active_group<-}
\alias{active_group_replace}
\alias{active_group<-.sftrack}
\alias{active_group<-.sftraj}
\alias{active_group<-.c_grouping}
\title{Set new active group}
\usage{
active_group(x) <- value

\method{active_group}{sftrack}(x) <- value

\method{active_group}{sftraj}(x) <- value

\method{active_group}{c_grouping}(x) <- value
}
\arguments{
\item{x}{sftrack/sftraj/c_grouping/s_group object}

\item{value}{character vector of the grouping names to make active}
}
\description{
Set new active group
}
