% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfstep.R
\name{as_sfstep}
\alias{as_sfstep}
\title{as_sfstep}
\usage{
as_sfstep(data, ...)
}
\arguments{
\item{data}{Data.frame input, these columns will remain unchanged, and any columns refered to
in later parameters are not deleted from this column. The function simply copies the data.frame
and adds the appropriate columns.}

\item{proj4}{projection (for sf)}

\item{time}{vector of time}

\item{burst}{list of named vectors one of which must be named id}

\item{error}{error vector}

\item{coords}{vector of three column names for the x,y,z coordinates, in that order.}

\item{tz}{timezone component, same as as.POSIX}
}
\description{
This generic has multiple inputs and gathers relevant information
to sfstep class.
It converts x,y,z data into an sfstep object and gives it an sf$geometry column. This
column is a list of line segments representing each step. It also creates and error, time, and burst column as well of each respective class.
}
