% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftrack.R
\name{as_sftrack}
\alias{as_sftrack}
\title{as_sftrack}
\usage{
as_sftrack(data, ...)
}
\arguments{
\item{data}{Data.frame input, these columns will remain unchanged, and any columns refered to
in later parameters are not deleted from this column. The function simply copies the data.frame
and adds the appropriate columns.}

\item{proj4}{projection (for sf)}

\item{time}{vector of time}

\item{burst}{list of named vectors one of which must be named id}

\item{error}{error vector}

\item{coords}{vector of three column names for the x,y,z coordinates, in that order.}

\item{tz}{timezone component, same as as.POSIX}
}
\description{
This generic has multiple inputs and gathers relevant information
to sftrack class.
It converts x,y,z data into an sftrack object and gives it an sf$geometry column. This
column is a list of points. It also creates and error, time, and burst column as well of each respective class.
}
\examples{
  # Input is a ltraj

  # Input is a sf object

  # Input is an sftrack object
}
