% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{burst-class}
\alias{burst-class}
\alias{make_ind_burst}
\alias{make_multi_burst}
\alias{c.ind_burst}
\alias{c.multi_burst}
\title{A class to group movement data}
\usage{
make_ind_burst(burst)

make_multi_burst(burst_list = NULL, active_burst = NULL)

\method{c}{ind_burst}(...)

\method{c}{multi_burst}(..., recursive = FALSE)
}
\arguments{
\item{burst}{a list containing named burst variables, one item must be named 'id'. ex: list(id = 1, month = 'may'). In the case of a multi_burst can be a list of ind_bursts.}

\item{burst_list}{a list of equal length vectors which will be combined to create a multi_burst. ex: list(x = 1st_vector, y = 2nd_vector)}

\item{active_burst}{a vector of the names of the bursts to be considered 'active' for the sake of analysis.}

\item{...}{objects to be pasted together into a multi_burst}

\item{recursive}{ignored}
}
\description{
This class describes a burst, which is a grouping variable
in which to split the data for analysis purposes.
It is composed of a list with named vectors.
One of which must be 'id', the id of the subject monitored,
and can be any length beyond that.
}
\details{
A burst is a list of possible grouping categories. The 'active_bursts'
of these groups is a collection of these variables that when combined form a singular group
for analysis purposes. The 'active_burst' can be any combination of the categories in a burst,
and can change with the use of `active_burst()`.

An `ind_burst` is a single rows burst. It is a 1xn dimensional list with any length(n) > 1.
Atleast one of the bursts must be named 'id' which is the subjects id.

A `multi_burst` is a collection of `ind_burst`s, it is a data.frame with dimensions of 1xnrow(data).
One multi_burst has one 'active_burst' which describes the set of names in each ind_burst to use as grouping
variables for analysis purposes. When you change the 'active_burst', calculations and plots change accordingly
to the new grouping levels.

You can create bursts with make_ind_burst and make multi_burst.
}
\examples{
# Make a single burst
make_ind_burst(burst=list(id='CJ11',month=3, height=10))

# Make a multi burst
 raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
 burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
 mb1 <- make_multi_burst(burst_list=burstz, active_burst=c('id','month'))
 str(mb1)

# Make a multi_burst from many ind_bursts
 a <- make_ind_burst(list(id = 1, year = 2020))
 b <- make_ind_burst(list(id = 1, year = 2020))
 c <- make_ind_burst(list(id = 2, year = 2020))

 c(a, b, c)
}
