% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{burst_levels}
\alias{burst_levels}
\alias{burst_levels<-}
\title{Access the burst levels}
\usage{
burst_levels(burst, value)

burst_levels(burst) <- value
}
\arguments{
\item{burst}{a multi_burst}

\item{value}{a character vector of the new levels, must contain all the current levels}
}
\description{
Access the burst levels
}
\examples{

raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
 mb1 <- make_multi_burst(burst_list = list(id = raccoon_data$sensor_code))
 # Access the burst levels
 burst_levels(mb1)

 # edit the burst levels by adding more values
 burst_levels(mb1) <- c('CJ11', 'CJ12', 'CJ13')
}
