% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{calc_sort_index}
\alias{calc_sort_index}
\title{#' @export
calc_burst_levels <- function(x){
  burst_levels <- lapply(names(x[[1]]), function(y) unique(sapply(x, function(z) z[[y]])))
  names(burst_levels) <- names(x[[1]])
  return(burst_levels)
}}
\usage{
calc_sort_index(burst, active_burst)
}
\description{
#' @export
calc_burst_levels <- function(x){
  burst_levels <- lapply(names(x[[1]]), function(y) unique(sapply(x, function(z) z[[y]])))
  names(burst_levels) <- names(x[[1]])
  return(burst_levels)
}
}
