% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_burst_names}
\alias{check_burst_names}
\title{Are burst names equivalent for each ind_burst?}
\usage{
check_burst_names(burst)
}
\arguments{
\item{burst}{a multi_burst}
}
\description{
Are burst names equivalent for each ind_burst?
}
