% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_group_names}
\alias{check_group_names}
\title{Are group names equivalent for each s_group?}
\usage{
check_group_names(x)
}
\arguments{
\item{x}{a c_grouping}
}
\description{
Are group names equivalent for each s_group?
}
