% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_NA_coords}
\alias{check_NA_coords}
\title{Check if coordinates contain NAs in some columns but not others}
\usage{
check_NA_coords(xyz)
}
\arguments{
\item{xyz}{a data.frame of xy or xyz coordinates}
}
\description{
Check if coordinates contain NAs in some columns but not others
}
