% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_names_exist}
\alias{check_names_exist}
\title{Checks if inputted column names are in data.frame}
\usage{
check_names_exist(data, burst_col, coords, error_col, time_col)
}
\description{
Checks if inputted column names are in data.frame
}
