% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ordered}
\alias{check_ordered}
\title{Checks if burst is ordered by time and then outputs the correct order}
\usage{
check_ordered(burst, time_data, return = TRUE)
}
\arguments{
\item{burst}{a multi_burst}

\item{time_data}{a vector of time}

\item{return}{the new order or not?}
}
\description{
Checks if burst is ordered by time and then outputs the correct order
}
