% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sort}
\alias{check_sort}
\title{Checks if sort_index needs to be recalculated then recalculates them}
\usage{
check_sort(burst)
}
\arguments{
\item{burst}{a multi_burst}
}
\description{
Checks if sort_index needs to be recalculated then recalculates them
}
