% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_t_regular}
\alias{check_t_regular}
\title{Check if time is regular for each burst and returns logical for each burst}
\usage{
check_t_regular(x)
}
\arguments{
\item{x}{an sftrack/sftraj object}
}
\description{
Check if time is regular for each burst and returns logical for each burst
}
