% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\name{as_sftrack.track_xy}
\alias{as_sftrack.track_xy}
\alias{sftrack_to_track_xy}
\title{convert to and from amt class 'track_xy'}
\usage{
\method{as_sftrack}{track_xy}(data, ...)

sftrack_to_track_xy(data)
}
\arguments{
\item{data}{the data either sftrack,sftraj,or track_xy class to convert}

\item{...}{ignored}
}
\description{
convert to and from amt class 'track_xy'
}
