% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraj.R
\name{as_sftrack.ltraj}
\alias{as_sftrack.ltraj}
\alias{as_sftraj.ltraj}
\alias{sftrack_to_ltraj}
\alias{sftraj_to_ltraj}
\title{convert to and from ltraj}
\usage{
\method{as_sftrack}{ltraj}(data, ...)

\method{as_sftraj}{ltraj}(data, ...)

sftrack_to_ltraj(data)

sftraj_to_ltraj(data)
}
\arguments{
\item{data}{the data either sftrack,sftraj,or ltraj class to convert}

\item{...}{ignored}
}
\description{
convert to and from ltraj

#' @name as
#' @rdname coerce-methods
#' @aliases coerce,sftrack
#' @export
setOldClass('ltraj')
setAs("sftrack", "ltraj",function(from){
  sftrack_to_ltraj(from)})

#' @name as
#' @rdname coerce-methods
#' @aliases coerce,sftracj
#' @export
setAs("sftraj", "ltraj",function(from){
  sftraj_to_ltraj(from)})
}
\examples{
#'
### Ltraj Methods
# Input is a ltraj
library(adehabitatLT)
ltraj_df <- as.ltraj(
  xy = raccoons[, c("longitude", "latitude")],
  date = raccoons$timestamp,
  id = raccoons$animal_id, typeII = TRUE,
  infolocs = raccoons[, 1:6]
)
my_sftrack <- as_sftrack(ltraj_df)
head(my_sftrack)

my_sftraj <- as_sftraj(ltraj_df)
head(my_sftraj)

# convert back
sftrack_to_ltraj(my_sftrack)
sftraj_to_ltraj(my_sftraj)
}
