% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{as_sftrack.trackeRdata}
\alias{as_sftrack.trackeRdata}
\alias{as_sftraj.trackeRdata}
\alias{sftrack_to_trackeR}
\alias{sftraj_to_trackeR}
\title{convert to and from trackeR}
\usage{
\method{as_sftrack}{trackeRdata}(data, ..., include_units = FALSE)

\method{as_sftraj}{trackeRdata}(data, ..., include_units = FALSE)

sftrack_to_trackeR(data, sport, ...)

sftraj_to_trackeR(data, sport, ...)
}
\arguments{
\item{data}{the data either sftrack,sftraj,or trackeRdata class to convert}

\item{...}{More arguments to be passed to `trackerData()`, please see its help file for more information.}

\item{include_units}{whether to add the units() to the ends of the colnames so they may be retained when converting to sftrack}

\item{sport}{(for amt conversions) The sport (cycling, running, swimming) to data is taken from}
}
\description{
convert to and from trackeR

#' @title Coerce methods
#' @name as
#' @rdname coerce-methods
#' @aliases coerce,sftrack
#' # setAs("sftrack", "trackeRdata",function(from){#   sftrack_to_trackeR(from,...)}
}
\examples{
#'

library(trackeR)
filepath <- system.file("extdata/tcx/", "2013-06-01-183220.TCX.gz", package = "trackeR")
runDF <- readTCX(file = filepath, timezone = "GMT")
runTr0 <- trackeRdata(runDF)

# convert to sftrack
my_sftrack <- as_sftrack(runTr0)
head(my_sftrack)
head(as_sftraj(runTr0))

# convert back
sftrack_to_trackeR(my_sftrack, sport = "running")
}
