% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{dup_timestamp}
\alias{dup_timestamp}
\title{check that time is unique}
\usage{
dup_timestamp(x, time)
}
\arguments{
\item{x}{An sftrack/sftraj object or a multi_burst}

\item{time}{vector of time, not required if given a sftrack object.}
}
\description{
check that time is unique
}
