% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{fix_zero}
\alias{fix_zero}
\title{fix 0's to NAs in latitude and longitude}
\usage{
fix_zero(xyz)
}
\arguments{
\item{xyz}{a data.frame of xy or xyz coordinates}
}
\value{
returns a data.frame with 0s replaced with NAs
}
\description{
fix 0's to NAs in latitude and longitude
}
