% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tj.R
\name{geom_sftrack}
\alias{geom_sftrack}
\alias{geom_sftrack.sftrack}
\alias{geom_sftrack.sftraj}
\title{Function to plot sftrack objects in ggplot}
\usage{
geom_sftrack(data, ...)

\method{geom_sftrack}{sftrack}(mapping = aes(), data = NULL, ...)

\method{geom_sftrack}{sftraj}(mapping = aes(), data = NULL, ...)
}
\arguments{
\item{data}{the sftraj or sftrack object.}

\item{...}{arguments to passed to ggplot}
}
\description{
This function can be added to ggplot() to plot an sftrack and sftraj
Function does not yet work with ggplot grammer so you must but data= in this function
}
\examples{
#'
library(ggplot2)
library(sftrack)
raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
raccoon_data$acquisition_time <- as.POSIXct(raccoon_data$acquisition_time, 'EST')
  burstz <- c(id = 'sensor_code')

# sftraj will as well for the most part, however as its a more complex structure to speed up plotting.
my_sftraj <- as_sftraj(raccoon_data, time = 'acquisition_time',
  coords = c('longitude','latitude'),
  burst = burstz)

ggplot() + geom_sftrack(data = my_sftraj)
}
