% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{group_labels}
\alias{group_labels}
\alias{group_labels.sftrack}
\alias{group_labels.sftraj}
\alias{group_labels.c_grouping}
\title{Shows grouping labels created from the s_group and the c_grouping}
\usage{
group_labels(x)

\method{group_labels}{sftrack}(x)

\method{group_labels}{sftraj}(x)

\method{group_labels}{c_grouping}(x)
}
\arguments{
\item{x}{a sftrack or grouping object}
}
\description{
Shows grouping labels created from the s_group and the c_grouping
}
