% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{group_names}
\alias{group_names}
\alias{group_name}
\alias{group_names.c_grouping}
\alias{group_names.sftraj}
\alias{group_names.sftrack}
\title{Display the levels of the sort index}
\usage{
group_names(x)

\method{group_names}{c_grouping}(x)

\method{group_names}{sftraj}(x)

\method{group_names}{sftrack}(x)
}
\arguments{
\item{x}{sftrack/sftraj/c_grouping/s_group object}
}
\description{
Display the levels of the sort index
}
