% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraj2sftrack.R
\name{ltraj2track}
\alias{ltraj2track}
\title{Convert ltraj to sftrack}
\usage{
ltraj2track(data)
}
\arguments{
\item{data}{ltraj object created from adehabitatLT}

\item{proj4}{projection (for sf)}

\item{tz}{timezone component, same as as.POSIX}
}
\description{
This function converts ltraj to sftrack by making some assumptions on data structure from the ltraj
}
\details{
Convert ltraj to sftrack
}
\examples{
library(adehabitatLT)
data(raccoon_data)
ltraj_df <- as.ltraj(xy=raccoon_data[,c('longitude','latitude')], date = as.POSIXct(raccoon_data$acquisition_time),
 id = raccoon_data$sensor_code, typeII = TRUE,
 infolocs = raccoon_data[,1:6] )
ltraj2track(data = ltraj_df)
}
