% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{make_ind_burst}
\alias{make_ind_burst}
\title{Burst Class}
\usage{
make_ind_burst(burst)
}
\arguments{
\item{burst}{burst}

\item{id}{id}

\item{newlevels}{a list of new levels if youd like to override the precalculated levels in a burst. ex. list(id = c('new','levels'))}
}
\description{
This class holds bursts, which is a place-holder name
that describes any grouping variable to split the data points, in which only the animal ID
is required, and otherwise can have n number of other grouping names.
Must be a named list
}
\details{
ind_burst()
multi_burst()
make_bulti_burst()
}
\examples{
# Make a single burst
make_ind_burst(burst=list(id='CJ11',month=3, height=10))

# Make a multi burst
 data(raccoon_data)
 burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
 mb1 <- make_multi_burst(burst_list=burstz, active_burst=c('id','month'))
 str(mb1)
}
