% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.R
\name{merge_traj}
\alias{merge_traj}
\title{Merge connected lines and create an sf object}
\usage{
merge_traj(x, mode = c("steps", "trajectories"))
}
\arguments{
\item{x}{an sftraj object}

\item{mode}{Character; either of \code{"steps"} (default), or
\code{"trajectories"}.}
}
\description{
This function returns a sf object grouped with a
    MULTILINESTRING for each grouping in a geometry column.
}
