% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_tj.R
\name{new_error_tj}
\alias{new_error_tj}
\title{Error Class
new_error_tj()}
\usage{
new_error_tj(error = NA, ...)
}
\arguments{
\item{error}{error}
}
\description{
This class exists as a placeholder. We do not
Currently know the details of what the error class
will look like.
}
\examples{
new_error_tj()
}
