% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftraj.R
\name{new_sftraj}
\alias{new_sftraj}
\title{sftraj Class}
\usage{
new_sftraj(data = data.frame(), proj4 = NA, time = NA, id = NA,
  burst = NULL, error = NA, coords = c("x", "y", "z"), tz = NULL)
}
\arguments{
\item{data}{Data.frame input, these columns will remain unchanged, and any columns refered to
in later parameters are not deleted from this column. The function simply copies the data.frame
and adds the appropriate columns.}

\item{proj4}{projection (for sf)}

\item{time}{vector of time}

\item{id}{vector of ids for the data}

\item{burst}{list of named vectors}

\item{error}{error vector}

\item{coords}{vector of three column names for the x,y,z coordinates, in that order.}

\item{tz}{timezone component, same as as.POSIX}
}
\description{
This is the highest level class that collects the error, time, and burst class.
It converts x,y,z data into an sftraj object and gives it an sf$geometry column,
and creates and error, time, and burst column as well of each respective class.
}
\examples{
 data(raccoon_data)
 burstz <- list(month = as.POSIXlt(raccoon_data$utc_date)$mon, height =as.numeric(raccoon_data$height>5))
my_traj <- new_sftraj(raccoon_data, time =as.POSIXct(raccoon_data$acquisition_time), id = raccoon_data$sensor_code,
  error = NA, coords = c('longitude','latitude','height'), tz = 'UTC',
  burst =burstz)
}
