% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ordered}
\alias{ordered}
\title{Checks if burst is ordered by time and then outputs the correct order}
\usage{
ordered(burst, time_data, return = TRUE)
}
\description{
Checks if burst is ordered by time and then outputs the correct order
}
