% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tj.R
\name{plot.sftraj}
\alias{plot.sftraj}
\title{methods for plot sftrack/sftraj}
\usage{
\method{plot}{sftraj}(x, ...)
}
\arguments{
\item{x}{sftrack object}

\item{...}{arguments to passed to plot}
}
\description{
methods for plot sftrack/sftraj
}
