% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tj.R
\name{plot_sftrack}
\alias{plot_sftrack}
\alias{plot.sftrack}
\alias{plot.sftraj}
\title{Methods for plotting sftrack/sftraj}
\usage{
\method{plot}{sftrack}(x, y, key.pos, key.width, ...)

\method{plot}{sftraj}(x, y, key.pos, key.width, ..., mode)
}
\arguments{
\item{x}{'sftrack' or 'sftraj' object}

\item{y}{ignored}

\item{key.pos}{Integer; side to plot a color key: 1 bottom, 2 left,
3 top, 4 right; set to NULL to omit key, or -1 to select
automatically (defaults to 4; see \code{\link[sf]{plot_sf}} for
more details).}

\item{key.width}{Amount of space reserved for the key, including
labels (see \code{\link[sf]{plot_sf}} for more details).)}

\item{...}{Further arguments passed to 'plot.sf'. Among others,
arguments for the key are set differently in 'sftrack' to allow
for longer labels by default (but can be nevertheless
adjusted).}

\item{mode}{Character; either of \code{"steps+points"},
\code{"steps"}, or \code{"trajectories"}, defaults to
\code{"steps+points"}, and switch to \code{"steps"} if there
are more than 10,000 steps.}
}
\description{
Methods for plotting sftrack/sftraj

Methods for plotting sftrack/sftraj
}
\details{
Plotting mode refers to considering the trajectory as
    connected elements. For \code{"steps+points"}, this means it
    will plot each step & point individually. The alternative,
    faster \code{"steps"} mode merges connected steps into
    multilinestrings, which are plotted as continuous lines.  This
    approach is much faster to plot with large objects, and is thus
    turned automatically on when n(steps)>10,000. Finally,
    \code{"trajectories"} merges the full trajectory as a continuous
    line (i.e. ignores gaps).
}
\examples{

## Plotting with sftrack is just like sf. `...` will accept most
## arguments as 'plot.sf':
plot(racc_track, axes = TRUE, lwd = 5, cex = 5, bgc = "gray50")

## sftraj will as well for the most part; however it is a more complex
## structure that combines points and steps (in step mode):
plot(racc_traj, lwd = 5, cex = 5, bgc = "gray50", graticule = TRUE)

}
