% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftraj.R
\name{Print_sftraj_objects}
\alias{Print_sftraj_objects}
\alias{print.sftraj}
\title{Print methods for sftraj}
\usage{
\method{print}{sftraj}(x, n = getOption("sf_max_print", default = 6), n_col, ...)
}
\arguments{
\item{x}{sftraj object}

\item{n}{Integer of number of records to display. Defaults to 6,
can be set globally by the \code{sf} option
\code{options(sf_max_print = ...)}}

\item{n_col}{Integer of number of columns to display + required sftrack columns (burst, geometry, time, and error). Defaults to global option default if non supplied}

\item{...}{other arguments passed onto print}
}
\description{
Print methods for sftraj
}
