% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{step_calc}
\alias{step_calc}
\title{Calculates step metrics including distance, dt, dx, and dy.}
\usage{
step_calc(sftrack)
}
\arguments{
\item{sftrack}{an sftrack/sftraj object}
}
\description{
Calculates step metrics including distance, dt, dx, and dy.
}
