% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{calc_dx}
\alias{calc_dx}
\alias{calc_dy}
\alias{calc_difftime}
\alias{calc_steplen}
\alias{calc_absangle}
\alias{calc_speed}
\alias{calc_relangle}
\title{step calculation functions}
\usage{
calc_dx(x)

calc_dy(x)

calc_difftime(x)

calc_steplen(x)

calc_absangle(x, ..., dx, dy, len)

calc_speed(..., x, len, dt)

calc_relangle(..., x, abs_angle)
}
\arguments{
\item{x}{an sftrack/sftraj object or in some cases a geometry}

\item{...}{ignored}

\item{dx}{(optional) output of calc_dx if already calculated. Otherwise will call `calc_dx()` internally}

\item{dy}{(optional) output of calc_dy if already calculated. Otherwise will call `calc_dy()` internally}

\item{len}{(optional) output of calc_steplen if already calculated. Otherwise will call `calc_steplen()` internally}

\item{dt}{(optional) output of calc_difftime if already calculated. Otherwise will call `calc_difftime()` internally}

\item{abs_angle}{(optional) output of calc_absangle if already calculated. Otherwise will call `calc_absangle()` internally}
}
\description{
step calculation functions
}
