% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{step_metrics}
\alias{step_metrics}
\title{Calculates step metrics including distance, dt, dx, and dy.}
\usage{
step_metrics(x)
}
\arguments{
\item{x}{an sftrack/sftraj object. sftrack objects will be converted to sftraj internally for calculation.}
}
\description{
Calculates step metrics including distance, dt, dx, and dy.
}
\examples{

step_metrics(racc_traj)[1:10, ]

}
