% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{step_recalc}
\alias{step_recalc}
\title{recalculate step geometry}
\usage{
step_recalc(x, return = FALSE)
}
\arguments{
\item{x}{an sftraj object.}

\item{return}{return step_geometry instead of replacing sftraj object with new step geometry. Defaults to FALSE}
}
\description{
Step geometeries in sftraj objects are linestrings going from t1 to t2 of a 'step'. As these are stored at the row level they are not dynamic to changes in t2.
step_recalc allows you to recalculate these geometeries if your data.frame has changed because of subsetting or filtering.
}
