% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tj.R
\name{t2_by_group}
\alias{t2_by_group}
\title{Calculate t2 based on groupings}
\usage{
t2_by_group(time, group)
}
\arguments{
\item{time}{time vector, either sft_timestamp, POSIXct, or numeric}

\item{group}{c_grouping object}
}
\description{
Calculate t2 based on groupings
}
