% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tj.R
\name{time-class}
\alias{time-class}
\alias{sft_time}
\title{A class to store time and time intervals}
\usage{
sft_time(x = list())
}
\arguments{
\item{x}{a list or vector containing time objects, either POSIXct or integer. If an sftraj must be a time interval where [[1]] ==start and [[2]] == end time for step.}
}
\description{
Time in movement data depends on if its a track (sftrack) or trajectory (sftraj).
A track has a single timestamp for each data point where as a trajectory has a start and end point for each step.
The sft_timestamp class is a column level class containing a list or vector for each point.
}
