% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tj.R
\name{time_recalc}
\alias{time_recalc}
\title{recalculate time objects from t1}
\usage{
time_recalc(x)
}
\arguments{
\item{x}{an sftraj object}
}
\description{
Time at time2 maybe recalculated due to subsetting or converting to new objects. This function is for sftraj objects.
If you need to recalculate a sft_timestamp on its own, you can use make_timestamp and feed it the new t1s.
}
