% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{active_burst<-}
\alias{active_burst<-}
\title{change active burst}
\usage{
active_burst(burst) <- value
}
\description{
change active burst
}
\examples{
mb <- ind_burst(list(id=1, group=2),active_burst='id')
active_burst(mb)
active_burst(mb) <- c('id','group')
active_burst(mb)
}
