% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{active_burst}
\alias{active_burst}
\title{Access the active_burst value}
\usage{
active_burst(burst)
}
\arguments{
\item{burst}{a multi_burst}
}
\description{
The active burst is the combination of bursts that group the data sets.
The active_bursts are essentially a paste(names_of_bursts, sep = '_') grouping variable.
}
\examples{
data('raccoon')
raccoon$acquisition_time <- as.POSIXct(raccoon$acquisition_time, 'EST')
 burstz <- list(id = raccoon$sensor_code,month = as.POSIXlt(raccoon$utc_date)$mon)
 mb1 <- make_multi_burst(x=burstz, active_burst=c('id','month'))

 # see the current active burst
 active_burst(mb1)

 # change the active burst
 active_burst(mb1) <- 'id'

 # Using a full data set
 my_track <- as_sftrack(raccoon, time = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst = burstz)

 summary(my_track)

 # change active_burst
 active_burst(my_track$burst) <- 'id'

 summary(my_track)
#'
}
