% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftrack.R
\name{as_sftrack}
\alias{as_sftrack}
\alias{new_sftrack}
\alias{as_sftrack.data.frame}
\alias{as_sftrack.sftraj}
\alias{as_sftrack.ltraj}
\alias{as_sftrack.sf}
\title{Convert objects into sftrack objects.}
\usage{
as_sftrack(data, ...)

new_sftrack(data, burst, time, geometry, error = NA)

\method{as_sftrack}{data.frame}(data, ..., xyz, coords = c("x", "y",
  "z"), burst_list, id, burst_col = NULL, active_burst = NA, error,
  error_col, time, time_col, crs = NA, zeroNA = F)

\method{as_sftrack}{sftraj}(data, ...)

\method{as_sftrack}{ltraj}(data, ..., crs = NA)

\method{as_sftrack}{sf}(data, ..., burst_list, id, burst_col = NULL,
  active_burst = NA, error, error_col, time, time_col)
}
\arguments{
\item{data}{(global) a data.frame of the movement data}

\item{...}{extra information to be past to as_sftrack}

\item{burst}{a multi_burst}

\item{time}{(vector) a vector of time information, can be either POSIX or an integer}

\item{geometry}{a vector of sf geometries}

\item{error}{(vector - optional) a vector of error information for the movement data}

\item{xyz}{(vector) a data.frame of xy or xyz coordinates}

\item{coords}{(data.frame) a character vector describing where the x,y,z (optional) coordinates are located in `data`}

\item{burst_list}{(vector) a list of named vectors describing multiple grouping variables}

\item{id}{(data.frame) a character string naming the 'id' field in `data`. Required if using data.frame inputs}

\item{burst_col}{(data.frame - optional) a character vector naming the other grouping columns in `data`.}

\item{active_burst}{(global) a character vector of the burst names to be 'active' to group data by for analysis}

\item{error_col}{(data.frame - optional) a character string naming the column in `data` where the error information is located}

\item{time_col}{(data.frame) a character string naming the column in `data` where the time information is located}

\item{crs}{a crs string from rgdal of the crs and projection information for the spatial data. Defaults to NA}

\item{zeroNA}{logical whether to convert 0s in spatial data into NAs. Defaults to FALSE.}
}
\description{
This function converts x,y,z data into an sftrack object with a sf_geometry column of sf_POINTS.
Creates a `burst` column to group movement data, and dedicated time and error columns.

Raw data input can be done using two 'modes': vector or data.frame. 'Vector' inputs gives the argument as a vector where
length = nrow(data). 'Data.frame' inputs gives the arguments as the column name of `data` where the input can be found.
Either input is allowed, but vector mode over.writes data.frame mode if both are given.

Some options are global and required regardless
}
\details{
Convert objects into sftrack objects.
}
\examples{

raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
  burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
  # Input is a data.frame
my_track <- as_sftrack(raccoon_data, time_col = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst_list = burstz)

  # Input is a ltraj
  library(adehabitatLT)
  ltraj_df <- as.ltraj(xy=raccoon_data[,c('longitude','latitude')],
  date = as.POSIXct(raccoon_data$acquisition_time),
  id = raccoon_data$sensor_code, typeII = TRUE,
  infolocs = raccoon_data[,1:6] )
  my_sftrack <- as_sftrack(ltraj_df)
  head(my_sftrack)

  # Input is a sf object
  library(sf)
  df1 <- raccoon_data[!is.na(raccoon_data$latitude),]
  sf_df <- st_as_sf(df1, coords=c('longitude','latitude'))
  id = 'sensor_code'
  time_col = 'acquisition_time'

  new_sftrack <- as_sftrack(sf_df, id=id, time_col = time_col)
  head(new_sftrack)

  # Input is an sftraj object
  my_traj <- as_sftraj(raccoon_data, time_col = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst_list = burstz)

  new_track <- as_sftrack(my_traj)
  head(new_track)
}
