% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftraj.R
\name{as_sftraj}
\alias{as_sftraj}
\alias{new_sftraj}
\alias{as_sftraj.data.frame}
\alias{as_sftraj.sftrack}
\alias{as_sftraj.sf}
\alias{as_sftraj.ltraj}
\title{Convert objects into sftraj objects.}
\usage{
as_sftraj(data = data.frame(), ...)

new_sftraj(data, burst_col, sf_col, time_col, error_col = NA)

\method{as_sftraj}{data.frame}(data = data.frame(), ...,
  coords = c("x", "y"), burst = "id", active_burst = NA,
  time = "time", error = NA, crs = NA, zeroNA = FALSE)

\method{as_sftraj}{sftrack}(data, ...)

\method{as_sftraj}{sf}(data, ..., coords, burst, active_burst = NA, time,
  error = NA)

\method{as_sftraj}{ltraj}(data, ...)
}
\arguments{
\item{data}{(global) a data.frame of the movement data}

\item{...}{extra information to be past to as_sftraj}

\item{burst_col}{(data.frame - optional) a character vector naming the other grouping columns in `data`.}

\item{time_col}{(data.frame) a character string naming the column in `data` where the time information is located}

\item{error_col}{(data.frame - optional) a character string naming the column in `data` where the error information is located}

\item{coords}{(data.frame) a character vector describing where the x,y,z (optional) coordinates are located in `data`}

\item{burst}{a multi_burst}

\item{active_burst}{(global) a character vector of the burst names to be 'active' to group data by for analysis}

\item{time}{(vector) a vector of time information, can be either POSIX or an integer}

\item{error}{(vector - optional) a vector of error information for the movement data}

\item{crs}{a crs string from rgdal of the crs and projection information for the spatial data. Defaults to NA}

\item{zeroNA}{logical whether to convert 0s in spatial data into NAs. Defaults to FALSE.}

\item{xyz}{(vector) a data.frame of xy or xyz coordinates}

\item{burst_list}{(vector) a list of named vectors describing multiple grouping variables}

\item{id}{(data.frame) a character string naming the 'id' field in `data`. Required if using data.frame inputs}

\item{geometry}{a vector of sf geometries}
}
\description{
This function converts x,y,z data into an sftrack object with a sf_geometry column of sf_POINTS.
Creates a `burst` column to group movement data and sets dedicated time and error columns.

Raw data input can be done using two 'modes': vector or data.frame. 'Vector' inputs gives the argument as a vector where
length = nrow(data). 'Data.frame' inputs gives the arguments as the column name of `data` where the input can be found.
Either input is allowed, but vector mode over.writes data.frame mode if both are given.

Some options are global and required regardless
}
\details{
Convert objects into sftraj objects.
}
\examples{

raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
raccoon_data$acquisition_time <- as.POSIXct(raccoon_data$acquisition_time, 'EST')
  burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
  # Input is a data.frame
my_traj <- as_sftraj(raccoon_data, time_col = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst_list = burstz)

  # Input is a ltraj
  library(adehabitatLT)
  ltraj_df <- as.ltraj(xy=raccoon_data[,c('longitude','latitude')],
  date = as.POSIXct(raccoon_data$acquisition_time),
  id = raccoon_data$sensor_code, typeII = TRUE,
  infolocs = raccoon_data[,1:6] )
  my_sftraj <- as_sftraj(ltraj_df)
  head(my_sftraj)

  # Input is a sf object
  library(sf)
  df1 <- raccoon_data[!is.na(raccoon_data$latitude),]
  sf_df <- st_as_sf(df1, coords=c('longitude','latitude'))
  id = 'sensor_code'
  time_col = 'acquisition_time'

  new_sftraj <- as_sftraj(sf_df, id=id, time_col = time_col)
  head(new_sftraj)

  # Input is an sftrack object
  my_track <- as_sftrack(raccoon_data, time_col = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst_list = burstz)

  new_traj <- as_sftraj(my_track)
  head(new_traj)
}
