% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{burst-class}
\alias{burst-class}
\alias{make_ind_burst}
\alias{make_multi_burst}
\alias{c.ind_burst}
\alias{c.multi_burst}
\title{A class to group movement data}
\usage{
make_ind_burst(x, active_burst = NULL)

make_multi_burst(x = NULL, active_burst = NULL)

\method{c}{ind_burst}(...)

\method{c}{multi_burst}(..., recursive = FALSE)
}
\arguments{
\item{x}{a list containing named burst variables, one item must be named 'id'. ex: list(id = 1, month = 'may'). For a multi_burst: A list of ind_bursts or
a list of equal length named vectors which will be combined to create a multi_burst. ex: list(x = 1st_vector, y = 2nd_vector)}

\item{active_burst}{a vector of the names of the bursts to be considered 'active' for the sake of analysis.}

\item{...}{objects to be pasted together into a multi_burst}

\item{recursive}{ignored}
}
\description{
This class describes a burst, which is a grouping variable for movement data.
It is composed of a list with named vectors.
One of which must be 'id', the id of the subject monitored,
and can be any length beyond that.
}
\details{
A burst is a list of possible grouping categories. The 'active_bursts'
of these groups is a collection of these variables that when combined form a singular group
for analysis purposes. The 'active_burst' can be any combination of the categories in a burst,
and can change with the use of `active_burst()`.

An `ind_burst` is a single rows burst. It is a 1xn dimensional list with any length(n) > 1.
Atleast one of the bursts must be named 'id' which is the subjects id.

A `multi_burst` is a collection of `ind_burst`s, it is a data.frame with dimensions of 1xnrow(data).
One multi_burst has one 'active_burst' which describes the set of names in each ind_burst to use as grouping
variables for analysis purposes. When you change the 'active_burst', calculations and plots change accordingly
to the new grouping levels.

You can create bursts with make_ind_burst and make multi_burst.
}
\examples{
# Make a single burst

make_ind_burst(x = list(id = "CJ11", month = 3, height = 10))

# Make a multi burst
data("raccoon")
raccoon$timestamp <- as.POSIXct(raccoon$timestamp, "EST")
burstz <- list(id = raccoon$animal_id, month = as.POSIXlt(raccoon$timestamp)$mon)
mb1 <- make_multi_burst(x = burstz, active_burst = c("id", "month"))
str(mb1)

# Make a multi_burst from many ind_bursts
a <- make_ind_burst(list(id = 1, year = 2020))
b <- make_ind_burst(list(id = 1, year = 2020))
c <- make_ind_burst(list(id = 2, year = 2020))

c(a, b, c)
}
