% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{burst_labels}
\alias{burst_labels}
\title{Shows burst labels created from the ind_burst and the active_burst}
\usage{
burst_labels(x, ...)
}
\arguments{
\item{x}{a sftrack or burst object}

\item{factor}{logical, whether to return a factor, default return is a character}
}
\description{
Shows burst labels created from the ind_burst and the active_burst
}
