% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{burst_labels}
\alias{burst_labels}
\title{Calculates burst labels created from the ind_burst and the active_burst}
\usage{
burst_labels(burst, factor = F, active_burst = NULL)
}
\arguments{
\item{burst}{a multi_burst}

\item{factor}{logical, whether to return a factor, defaults return is a character}

\item{active_burst}{(optional), the active_burst to subset by, defaults to the current active_burst.}
}
\description{
Calculates burst labels created from the ind_burst and the active_burst
}
