% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{burst_select}
\alias{burst_select}
\title{Select the active multi_bursts.}
\usage{
burst_select(burst, select = NULL)
}
\arguments{
\item{burst}{a multi_burst}

\item{select}{(optional), the burst names to subset by, defaults to the current active_burst.}
}
\value{
a list where each position is a list of the active bursts from each ind_burst
}
\description{
Select the active multi_bursts.
}
