% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{calc_sort_index}
\alias{calc_sort_index}
\title{Calculate a new sort index for bursts}
\usage{
calc_sort_index(x, active_burst = NA)
}
\arguments{
\item{x}{burst or sftrack object}

\item{active_burst}{(optional), a new active burst to calculate the sort index on. If not included, defaults  to the active burst (if a multi_burst) or the burst names}
}
\description{
Calculate a new sort index for bursts
}
