% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{calc_sort_index}
\alias{calc_sort_index}
\title{Calculate the sort index. Internal function}
\usage{
calc_sort_index(burst, active_burst)
}
\arguments{
\item{burst}{a multi_burst}

\item{active_burst}{a character vector of the active_burst}
}
\description{
Calculate the sort index. Internal function
}
