% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{calc_sort_index}
\alias{calc_sort_index}
\title{Calculate a new sort index for groups}
\usage{
calc_sort_index(x, active_group = NA)
}
\arguments{
\item{x}{group or sftrack object}

\item{active_group}{(optional), a new active group. If not included, defaults  to the active group (if a c_grouping) or the group names}
}
\description{
Calculate a new sort index for groups
}
