% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_names_exist}
\alias{check_names_exist}
\title{Check if a set of column names are found in a data frame and return an error if not}
\usage{
check_names_exist(data, names)
}
\arguments{
\item{data}{a data.frame to check names against}

\item{names}{the inputted column names}
}
\description{
Check if a set of column names are found in a data frame and return an error if not
}
