% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ordered}
\alias{check_ordered}
\title{Checks if grouping is ordered by time and then outputs the correct order}
\usage{
check_ordered(group, time_data, return = TRUE)
}
\arguments{
\item{group}{a c_grouping}

\item{time_data}{a vector of time}

\item{return}{T/F return the new order or just run check?}
}
\description{
Checks if grouping is ordered by time and then outputs the correct order
}
