% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\name{as_sftrack.track_xy}
\alias{as_sftrack.track_xy}
\alias{as_sftraj.track_xy}
\alias{sftrack_to_track_xy}
\title{convert to and from amt class 'track_xy'}
\usage{
\method{as_sftrack}{track_xy}(data, crs = attr(data, "crs_"), ...)

\method{as_sftraj}{track_xy}(data, ...)

sftrack_to_track_xy(data, ...)
}
\arguments{
\item{data}{the sftrack object}

\item{crs}{a crs string from rgdal of the crs and projection information for the spatial data. Defaults to the one provided by \code{data}.}

\item{...}{any arguments to be passed on to `amt::make_track()`. See the help file in `amt` for more details}
}
\description{
convert to and from amt class 'track_xy'
}
\examples{

library("amt")
data(sh)

## Add timestamp column, remove duplicated timestamps, and add ID
#column:
sh$timestamp <- as.POSIXct(paste(sh$day,sh$time))
sh <- sh[!duplicated(sh$timestamp), ]
sh$id <- "Animal 1"

## Create the 'track_xyt' object:
tr1 <- make_track(sh, x_epsg31467, y_epsg31467, timestamp,
                  id = id
)

head(as_sftrack(tr1, crs = "+init=epsg:31467"))
}
