% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.R
\name{group_summary}
\alias{group_summary}
\title{Summarize sftrack objects}
\usage{
group_summary(x, spatial = TRUE)
}
\arguments{
\item{x}{an \code{sftrack} or \code{sftraj} object}

\item{spatial}{Logical; whether to compute spatial statistics}
}
\description{
Summarize sftrack objects
}
