% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_tj.R
\name{ind_burst}
\alias{ind_burst}
\title{Burst Class}
\usage{
ind_burst(id = c(), burst = NULL, ...)
}
\arguments{
\item{burst}{}
}
\description{
This class holds bursts, which is a place-holder name
that describes any grouping variable to split the data points, in which only the animal ID
is required, and otherwise can have n number of other grouping names.
Must be a named list
}
\details{
ind_burst()
multi_burst()
make_bulti_burst()
}
\examples{
 data(raccoon_data)
 burstz <- list(month = as.POSIXlt(raccoon_data$utc_date)$mon, height =as.numeric(raccoon_data$height>5))
 make_multi_burst(id=raccoon_data$sensor_code,burst=burstz)

}
