% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{make_step_geom}
\alias{make_step_geom}
\title{Calculate step geometries given a set of bursts, time, and geometries}
\usage{
make_step_geom(burst, time_data, geometry)
}
\arguments{
\item{burst}{a multi_burst}

\item{time_data}{time vector}

\item{geometry}{the geometery data from either sf or sf_track. Must be an sf geometry class}
}
\description{
This calculates step geometries as individual line segments based on the active_burst
}
\examples{
raccoon_data <- read.csv(system.file('extdata/raccoon_data.csv', package='sftrack'))
raccoon_data$acquisition_time <- as.POSIXct(raccoon_data$acquisition_time, 'EST')
burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
data_sf <- as_sftrack(raccoon_data, time_col = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst_list =burstz)

make_step_geom(burst = data_sf$burst,
               geometry = data_sf$geometry,
               time_data = data_sf$acquisition_time)
}
