% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{make_step_geom}
\alias{make_step_geom}
\title{Calculate step geometries given a set of bursts, time, and geometries}
\usage{
make_step_geom(burst, time_data, geometry)
}
\arguments{
\item{burst}{a multi_burst object}

\item{time_data}{time vector}

\item{geometry}{the geometery data from either sf or sf_track. Must be an sf geometry class}
}
\description{
This calculates step geometries as individual line segments based on the active_burst
}
\examples{
library(sf)
geom = st_as_sf(data.frame(
  x = c(1,2,2,5),
  y = c(0,1,5,7)
), coords = c('x','y'))

burst = list(id = rep(1,4))
time = 1:4

mb = make_multi_burst(burst)

make_step_geom(burst = mb,
               geometry = geom$geometry,
               time_data = time)

}
