% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{make_step_geom}
\alias{make_step_geom}
\title{step geom calculator}
\usage{
make_step_geom(burst_id, time_data, geometry)
}
\arguments{
\item{burst_id}{burst object thats subsetted by only the bursts we want to use to calculate one step_geom}

\item{geometry}{the geometery data from either sf or sf_track. Must be an sf geometry class}

\item{timez}{time object}
}
\description{
This calculates step geom as individual line segments. It implies your bursts
}
\examples{
burstz <- list(id = raccoon_data$sensor_code,month = as.POSIXlt(raccoon_data$utc_date)$mon)
data_sf <- as_sftrack(raccoon_data, time_col = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'),
  burst_list =burstz)

make_step_geom(burst_id = data_sf$burst, geometry = data_sf$geometry, time_data = data_sf$acquisition_time)
}
