% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tj.R
\name{make_timestamp}
\alias{make_timestamp}
\title{Make timestamp object from a vector of time}
\usage{
make_timestamp(t1, group)
}
\arguments{
\item{t1}{Vector of the time at start of each point}

\item{group}{The sft_group object for grouping data.}
}
\description{
Make timestamp object from a vector of time
}
