% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.R
\name{merge_traj}
\alias{merge_traj}
\title{Merge connected lines and create an sf object}
\usage{
merge_traj(x)
}
\arguments{
\item{x}{an sftraj object}
}
\description{
This function returns a sf object grouped by each burst with a geometry column of multilinestrings for each grouping
}
