% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftraj.R
\name{new_sftraj}
\alias{new_sftraj}
\title{Define an sftraj}
\usage{
new_sftraj(data, group_col, sf_col, time_col, error_col = NA)
}
\arguments{
\item{data}{data.frame with multi_burst column, geometry column, time_col (integer/POSIXct), and error column (optional)}

\item{group_col}{column name of multi_burst in `data`}

\item{sf_col}{column name of geometry in `data`}

\item{time_col}{column name of time in `data`}

\item{error_col}{column name of error in `data`}
}
\description{
Define an sftraj
}
