% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tj.R
\name{new_time_tj}
\alias{new_time_tj}
\title{Methods for time traj class
new_time_tj()}
\usage{
new_time_tj(time = c(), id = c(), ..., tz = "UTC")
}
\arguments{
\item{time}{vector of time or as.POSIX format object.}

\item{tz}{as.POSIX time zone. Not necessary if an integer vector}
}
\description{
This class is a time object that can either be an integer vector or an as.POSIXct
object, but with extra checks that the time is unique and ordered.
}
\examples{
data(raccoon_data)
}
