% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tj.R
\name{plot_sftrack}
\alias{plot_sftrack}
\alias{plot.sftrack}
\alias{plot.sftraj}
\title{methods for plot sftrack/sftraj}
\usage{
\method{plot}{sftrack}(x, ...)

\method{plot}{sftraj}(x, ...)
}
\arguments{
\item{x}{sftrack/sftraj object}

\item{...}{arguments to passed to plot}
}
\description{
methods for plot sftrack/sftraj

methods for plot sftrack/sftraj
}
\examples{
library(sftrack)
data('raccoon')
raccoon$acquisition_time <- as.POSIXct(raccoon$acquisition_time, 'EST')
  burstz <- c(id = 'sensor_code')
my_sftrack <- as_sftrack(raccoon, time = 'acquisition_time',
   coords = c('longitude','latitude'),
  burst = burstz)

# Plotting with sftrack is just like sf. `...` will accept most arguments as plot.sf

plot(my_sftrack, axes = TRUE, lwd =5 ,cex=5, bgc = 'gray80')

# sftraj will as well for the most part, however as its a more
# complex structure to speed up plotting.
my_sftraj <- as_sftraj(raccoon, time = 'acquisition_time',
  coords = c('longitude','latitude'),
  burst = burstz)
plot(my_sftraj, axes = TRUE, lwd =5 ,cex=5, bgc = 'gray80', graticule = TRUE)
}
