% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftraj.R
\name{Print_sftraj_objects}
\alias{Print_sftraj_objects}
\alias{print.sftraj}
\title{Print methods for sftraj}
\usage{
\method{print}{sftraj}(x, n_row, n_col, ...)
}
\arguments{
\item{x}{sftraj object}

\item{n_row}{Integer of number of rows to display. Defaults to global option default if non supplied}

\item{n_col}{Integer of number of columns to display + required sftrack columns (burst, geometry, time, and error). Defaults to global option default if non supplied}

\item{...}{other arguments passed onto print}
}
\description{
Print methods for sftraj
}
