% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_calculation.R
\name{step_metrics}
\alias{step_metrics}
\title{Calculates step metrics including distance, dt, dx, and dy.}
\usage{
step_metrics(sftraj)
}
\arguments{
\item{sftraj}{an sftrack/sftraj object. sftrack objects will be converted to sftraj internally for calculation.}
}
\description{
Calculates step metrics including distance, dt, dx, and dy.
}
\examples{
#'
data('raccoon')
raccoon$acquisition_time <- as.POSIXct(raccoon$acquisition_time, 'EST')
  burstz <- list(id = raccoon$sensor_code,month = as.POSIXlt(raccoon$utc_date)$mon)
  # Input is a data.frame
my_sftraj <- as_sftraj(raccoon, burst = burstz, time = 'acquisition_time',
  error = NA, coords = c('longitude','latitude'))

step_metrics(my_sftraj)[1:10,]
}
