% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.R
\name{which_duplicated}
\alias{which_duplicated}
\title{Which grouping/time stamp combos are duplicated.}
\usage{
which_duplicated(data = data.frame(), group, time)
}
\arguments{
\item{data}{a data.frame containing burst or time data (if necessary)}

\item{group}{a list where each entry is a vector of groupings where length == nrow(data)|nrow(time). Or a character vector describing the column name they are located in data}

\item{time}{a vector of as.POSIXct time, or a character of the column name where it can be found in data}
}
\description{
This function returns a data.frame of which rows are duplicated and their time stamps.
}
